# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for HLTV-Manager.",
        "title": "HLTV-Manager backend",
        "contact": {},
        "version": "beta1.0"
    },
    "host": "0.0.0.0:3030",
    "basePath": "/",
    "paths": {
        "/api/v1/hltv": {
            "get": {
                "description": "Returns all configured HLTV instances with current runtime status and demo count.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hltv"
                ],
                "summary": "Get list of HLTV instances",
                "responses": {
                    "200": {
                        "description": "List of HLTV instances",
                        "schema": {
                            "$ref": "#/definitions/api.HLTVListResponse"
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/hltv/{id}": {
            "get": {
                "description": "Returns one HLTV instance by ID with current runtime status and demos count.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hltv"
                ],
                "summary": "Get HLTV instance details",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "HLTV ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "HLTV instance summary",
                        "schema": {
                            "$ref": "#/definitions/api.HLTVInfoResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid HLTV ID",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "HLTV not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "api.HLTVInfoResponse": {
            "type": "object",
            "properties": {
                "connect": {
                    "type": "string"
                },
                "demos_count": {
                    "type": "integer"
                },
                "game_id": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "port": {
                    "type": "string"
                },
                "running": {
                    "type": "boolean"
                },
                "show_ip": {
                    "type": "string"
                }
            }
        },
        "api.HLTVListItem": {
            "type": "object",
            "properties": {
                "connect": {
                    "type": "string"
                },
                "demos_count": {
                    "type": "integer"
                },
                "game_id": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "port": {
                    "type": "string"
                },
                "running": {
                    "type": "boolean"
                },
                "show_ip": {
                    "type": "string"
                }
            }
        },
        "api.HLTVListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.HLTVListItem"
                    }
                }
            }
        }
    }
}